refactor(core): move fetch-comments into packages/core
Relocates the headless fetch-comments orchestrator and its test. It imports only
fs, path, packages/core and the shared types, so it was already a Node program;
only its location said otherwise.

It travels separately from the rest of the move because it is separable on
different grounds. The rest moved so a second front end could exist. Nothing
consumes this module, and the serve program never will: the two are parallel
orchestrators over the same primitives rather than caller and callee. A reviewer
should be able to accept or reject that argument on its own.

This does NOT resolve issue #143. The desktop binary still imports and routes
this module, so Electron still initialises its platform, and that initialisation
is the crash. The move only puts the code where a command-line entry point could
later be attached; it does not attach one.
